MAINT add pyproject.toml, reduce role of setup.py
Adding pyproject.toml to be compliant with PEP 518.
For which build system to adopt, the competitors
are setuptools (current system in pylangacq), poetry, and flit.
It looks like packaging is (always?) in flux,
and that setuptools still has a semi-official status
(e.g., by its inclusion in pip) that poetry and flit don't have,
I'm reluctant to move away from setuptools at this point
(December 2021).
This being said, I've decided to specify build metadta and options
in setup.cfg fully and only keep a minimal setup.py around for
backward compatibility. Two reasons for this move:
* If I decide to move away from setuptools in the future,
  having all the details already specified in setup.cfg in a config
  language (though ini, not toml) should make things easier.
* Avoid setup.py as much as possible for all the good reasons,
  e.g., executing arbitrary code.

Squashed commit of the following:

commit 418022e36a329034e1ee583bdf166af7dc16a7a3
Author: jacksonllee <jacksonlunlee@gmail.com>
Date:   Sun Dec 26 22:05:34 2021 -0500

    MAINT update changelog

commit 44c5364c16c519bd25ce26cc0be72e3cff8c21ae
Author: jacksonllee <jacksonlunlee@gmail.com>
Date:   Sun Dec 26 21:55:24 2021 -0500

    MAINT add changelog url to setup.cfg

commit baa9a61848c7d2eff4b6e05277e13a405888f990
Author: jacksonllee <jacksonlunlee@gmail.com>
Date:   Sun Dec 26 20:12:34 2021 -0500

    MAINT minimal setup.py

commit 5ad5a6dec66caed59825c770459b89a7f7428669
Author: jacksonllee <jacksonlunlee@gmail.com>
Date:   Sun Dec 26 17:37:23 2021 -0500

    MAINT use importlib.metadata for version number